+++ orbit-core/src/commonMain/kotlin/org/orbitmvi/orbit/syntax/simple/SimpleSyntax.kt
@@ -21,7 +21,7 @@
 package org.orbitmvi.orbit.syntax.simple
 
 import org.orbitmvi.orbit.syntax.ContainerContext
-import org.orbitmvi.orbit.syntax.OrbitDsl
+import org.orbitmvi.orbit.annotation.OrbitDsl
 
 @OrbitDsl
 public class SimpleSyntax<S : Any, SE : Any>(public val containerContext: ContainerContext<S, SE>) {
