+++ orbit-core/src/commonMain/kotlin/org/orbitmvi/orbit/syntax/simple/SimpleSyntaxExtensions.kt
@@ -29,7 +29,7 @@ import org.orbitmvi.orbit.Container
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.idling.withIdling
 import org.orbitmvi.orbit.internal.runBlocking
-import org.orbitmvi.orbit.syntax.OrbitDsl
+import org.orbitmvi.orbit.annotation.OrbitDsl
 
 /**
  * Reducers reduce the current state and incoming events to produce a new state.
