+++ orbit-test2/src/commonMain/kotlin/org/orbitmvi/orbit/test/LiveTestSettingsBuilder.kt
@@ -7,38 +7,14 @@ import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import org.orbitmvi.orbit.RealSettings
 
 @OptIn(ExperimentalCoroutinesApi::class)
-public class TestSettingsBuilder2 internal constructor(
+public class LiveTestSettingsBuilder internal constructor(
     private var settings: RealSettings
 ) {
     public constructor() : this(
         UnconfinedTestDispatcher().let {
             RealSettings(
-                eventLoopDispatcher = it,
-                intentLaunchingDispatcher = it
-            )
-        }
-    )
-
-    public var exceptionHandler: CoroutineExceptionHandler?
-        get() = settings.exceptionHandler
-        public set(value) {
-            settings = settings.copy(exceptionHandler = value)
-        }
-
-    public var isolateFlow: Boolean = false
-
-    public fun build(): RealSettings = settings
-}
-
-@OptIn(ExperimentalCoroutinesApi::class)
-public class LiveTestSettingsBuilder2 internal constructor(
-    private var settings: RealSettings
-) {
-    public constructor() : this(
-        UnconfinedTestDispatcher().let {
-            RealSettings(
-                eventLoopDispatcher = it,
-                intentLaunchingDispatcher = it
+                    eventLoopDispatcher = it,
+                    intentLaunchingDispatcher = it
             )
         }
     )
